requirements.txt
# Updated requirements with more recent compatible versions
langchain>=0.1.0
langchain-community>=0.0.13
langchain-core>=0.1.0
langchain-experimental>=0.0.47
langchain-neo4j>=0.1.0
langchain-openai>=0.0.2
neo4j>=5.15.0
openai>=1.6.1
python-dotenv>=1.0.0
pydantic>=2.5.2
faiss-cpu>=1.7.4
numpy>=1.24.3
tiktoken>=0.5.2